requirements.txt
aiofiles==0.7.0
altgraph==0.17.2
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
decorator==5.1.0
fastapi==0.68.1
future==0.18.2
h11==0.12.0
idna==3.2
infi.systray==0.1.12
ipython==7.28.0
jedi==0.18.0
keyboard==0.13.5
matplotlib-inline==0.1.3
parso==0.8.2
pefile==2021.9.3
pickleshare==0.7.5
prompt-toolkit==3.0.20
pycodestyle==2.7.0
pydantic==1.8.2
Pygments==2.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pywin32-ctypes==0.2.0
requests==2.26.0
sortedcontainers==2.4.0
starlette==0.14.2
toml==0.10.2
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
wcwidth==0.2.5
wincertstore==0.2
